Implement more graphs
For now, have them displayed at the bottom so I can finish implementing
them all and consider renaming the files from camel case to lowercase
with hyphens. I think many of the old graphs, especially the ones that
offer extremely little display should be removed because the amount of
graphs is too insane for someone to spend their time looking through.
That or they need to be categorized. Just my thoughts.